move tests in api section to avoid mix with unit tests
